Safari-609.1.11.1/LayoutTests/fast/dom/NodeList/no-named-getter-expected.txt

Tests that NodeList does not have a named property getter.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS nodeList.__proto__ is NodeList.prototype
PASS nodeList['testId'] is undefined.
PASS nodeList.testId is undefined.
PASS nodeList['testName'] is undefined.
PASS nodeList.testName is undefined.
PASS successfullyParsed is true

TEST COMPLETE
